# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API",
        "title": "Dendrochronology API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/constructions/{id}/add-to-dendrochronology": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Добавляет конструкцию в черновик заявки; при первом добавлении создаётся заявка (201 + Location).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dendrochronologies"
                ],
                "summary": "Добавить в корзину",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID конструкции",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.DendrochronologyListJSON"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/serializer.DendrochronologyListJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/dendrochronology-constructions/{construction_id}/{dendrochronology_id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Образцы, даты резки и коррекции для пары конструкция–заявка",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dendrochronologies"
                ],
                "summary": "Изменить строку заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID конструкции",
                        "name": "construction_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "dendrochronology_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Поля",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializer.DendroConstructionUpdateJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.DendroConstructionUpdateJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dendrochronologies"
                ],
                "summary": "Удалить строку из заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID конструкции",
                        "name": "construction_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "dendrochronology_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.DendrochronologyListJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "serializer.DendroConstructionUpdateJSON": {
            "type": "object",
            "properties": {
                "cutting_date": {
                    "type": "string"
                },
                "date_correction": {
                    "type": "string"
                },
                "samples_count": {
                    "type": "integer"
                },
                "use_life": {
                    "type": "string"
                }
            }
        },
...
        "serializer.DendrochronologyListJSON": {
            "type": "object",
            "properties": {
                "build_date": {
                    "type": "integer"
                },
                "creator_login": {
                    "type": "string"
                },
                "date_completed": {
                    "type": "string"
                },
                "date_create": {
                    "type": "string"
                },
                "date_formed": {
                    "type": "string"
                },
                "dated_constructions_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "moderator_login": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_samples": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}